Move domain DI to app and remove Koin from domain
Relocate domain DI modules into the app module and update their package/imports to es.pedrazamiguez.splittrip.di.domain. Adjust Feature/Notification module aggregations to import the new package locations. Also remove Koin implementation/test dependencies from the domain module's build.gradle.kts to decouple the domain module from DI framework.
