+++ app/src/main/kotlin/es/pedrazamiguez/splittrip/di/FeatureModuleAggregations.kt
@@ -7,16 +7,16 @@ import es.pedrazamiguez.splittrip.data.di.groupsDataModule
 import es.pedrazamiguez.splittrip.data.di.profileDataModule
 import es.pedrazamiguez.splittrip.data.di.settingsDataModule
 import es.pedrazamiguez.splittrip.data.di.subunitsDataModule
-import es.pedrazamiguez.splittrip.domain.di.authenticationDomainModule
-import es.pedrazamiguez.splittrip.domain.di.balancesDomainModule
-import es.pedrazamiguez.splittrip.domain.di.contributionsDomainModule
-import es.pedrazamiguez.splittrip.domain.di.currenciesDomainModule
-import es.pedrazamiguez.splittrip.domain.di.expensesDomainModule
-import es.pedrazamiguez.splittrip.domain.di.groupsDomainModule
-import es.pedrazamiguez.splittrip.domain.di.profileDomainModule
-import es.pedrazamiguez.splittrip.domain.di.settingsDomainModule
-import es.pedrazamiguez.splittrip.domain.di.subunitsDomainModule
-import es.pedrazamiguez.splittrip.domain.di.withdrawalsDomainModule
+import es.pedrazamiguez.splittrip.di.domain.authenticationDomainModule
+import es.pedrazamiguez.splittrip.di.domain.balancesDomainModule
+import es.pedrazamiguez.splittrip.di.domain.contributionsDomainModule
+import es.pedrazamiguez.splittrip.di.domain.currenciesDomainModule
+import es.pedrazamiguez.splittrip.di.domain.expensesDomainModule
+import es.pedrazamiguez.splittrip.di.domain.groupsDomainModule
+import es.pedrazamiguez.splittrip.di.domain.profileDomainModule
+import es.pedrazamiguez.splittrip.di.domain.settingsDomainModule
+import es.pedrazamiguez.splittrip.di.domain.subunitsDomainModule
+import es.pedrazamiguez.splittrip.di.domain.withdrawalsDomainModule
 import es.pedrazamiguez.splittrip.features.authentication.di.authenticationUiModule
 import es.pedrazamiguez.splittrip.features.balance.di.balancesUiModule
 import es.pedrazamiguez.splittrip.features.contribution.di.contributionsUiModule
