+++ app/src/main/kotlin/es/pedrazamiguez/splittrip/di/NotificationModuleAggregations.kt
@@ -1,7 +1,7 @@
 package es.pedrazamiguez.splittrip.di
 
 import es.pedrazamiguez.splittrip.data.di.notificationsDataModule
-import es.pedrazamiguez.splittrip.domain.di.notificationsDomainModule
+import es.pedrazamiguez.splittrip.di.domain.notificationsDomainModule
 import org.koin.dsl.module
 
 val notificationModules = module {
