+++ app/src/main/kotlin/es/pedrazamiguez/splittrip/di/domain/CurrenciesDomainModule.kt
@@ -1,4 +1,4 @@
-package es.pedrazamiguez.splittrip.domain.di
+package es.pedrazamiguez.splittrip.di.domain
 
 import es.pedrazamiguez.splittrip.domain.repository.CurrencyRepository
 import es.pedrazamiguez.splittrip.domain.usecase.currency.GetExchangeRateUseCase
