+++ app/src/main/kotlin/es/pedrazamiguez/splittrip/di/domain/NotificationsDomainModule.kt
@@ -1,4 +1,4 @@
-package es.pedrazamiguez.splittrip.domain.di
+package es.pedrazamiguez.splittrip.di.domain
 
 import es.pedrazamiguez.splittrip.domain.repository.DeviceRepository
 import es.pedrazamiguez.splittrip.domain.repository.NotificationPreferencesRepository
